% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno.R, R/mafAnno_fast.R
\name{annotate_maf_with_facets_cf_tcn_lcn}
\alias{annotate_maf_with_facets_cf_tcn_lcn}
\title{don't know}
\usage{
annotate_maf_with_facets_cf_tcn_lcn(maf, out, fit,
  iTumor_Sample_Barcode = NULL)

annotate_maf_with_facets_cf_tcn_lcn(maf, out, fit,
  iTumor_Sample_Barcode = NULL)
}
\arguments{
\item{maf}{numeric first argument}

\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{iTumor_Sample_Barcode}{numeric first argument}

\item{diplogr}{numeric first argument}

\item{maf}{numeric first argument}

\item{fit}{numeric first argument}

\item{iTumor_Sample_Barcode}{numeric first argument}
}
\value{
don't know

don't know
}
\description{
don't know

temporary function
}
