% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno_fast.R
\name{annotate_maf_with_facets_cf_tcn_lcn_cncf}
\alias{annotate_maf_with_facets_cf_tcn_lcn_cncf}
\title{don't know}
\usage{
annotate_maf_with_facets_cf_tcn_lcn_cncf(maf, cncf)
}
\arguments{
\item{maf}{numeric first argument}

\item{cncf}{numeric first argument}
}
\value{
don't know
}
\description{
temporary function
}
