% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm-level-changes.R
\name{arm_level_changes}
\alias{arm_level_changes}
\title{Arm-level changes}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/29622463}
}
\usage{
arm_level_changes(segs, ploidy, genome = c("hg19", "hg18", "hg38"),
  algorithm = c("em", "cncf"))
}
\arguments{
\item{segs}{FACETS segmentation output.}

\item{ploidy}{Sample ploidy.}

\item{genome}{Genome build.}

\item{algorithm}{Choice between FACETS \code{em} and \code{cncf} algorithm.}
}
\value{
List with one or more values from function.
}
\description{
Get the altered chromosome arms in sample. Does not include the acrocentric p arms of chromosomes 12, 14, 15, 31, and 22.
}
