% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno_fast.R
\name{ccf.likelihood.purity.fit}
\alias{ccf.likelihood.purity.fit}
\title{don't know}
\usage{
ccf.likelihood.purity.fit(purity, absCN, alt_allele, coverage, copies)
}
\arguments{
\item{purity}{numeric first argument}

\item{absCN}{numeric first argument}

\item{alt_allele}{numeric first argument}

\item{coverage}{numeric first argument}

\item{copies}{numeric first argument}
}
\value{
don't know
}
\description{
temporary function
}
