% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno.R, R/mafAnno_fast.R
\name{ccf.likelihood}
\alias{ccf.likelihood}
\title{don't know}
\usage{
ccf.likelihood(purity, absCN, alt_allele, coverage, copies)

ccf.likelihood(purity, absCN, alt_allele, coverage, copies)
}
\arguments{
\item{purity}{numeric first argument}

\item{absCN}{numeric first argument}

\item{alt_allele}{numeric first argument}

\item{coverage}{numeric first argument}

\item{copies}{numeric first argument}

\item{purity}{numeric first argument}

\item{absCN}{numeric first argument}

\item{alt_allele}{numeric first argument}

\item{coverage}{numeric first argument}

\item{copies}{numeric first argument}
}
\value{
don't know

don't know
}
\description{
don't know

temporary function
}
