% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf-annotate-maf.R
\name{ccf_annotate_maf}
\alias{ccf_annotate_maf}
\title{Estimate CCFs of somatic mutations}
\format{Output columns:
\describe{
  \item{ccf_Mcopies*}{Inferred CCF if mutation is on the major allele.}
  \item{ccf_1copy*}{Inferred CCF if mutation exists in one copy.}
  \item{ccf_expected_copies*}{Inferred CCF if mutation exists in number of copies expected from observed VAF and local ploidy.}
}}
\usage{
ccf_annotate_maf(maf, segs, purity, algorithm = c("em", "cncf"))
}
\arguments{
\item{maf}{Input MAF file.}

\item{segs}{FACETS segmentation output.}

\item{purity}{Sample purity estimate.}

\item{algorithm}{Choice between FACETS \code{em} and \code{cncf} algorithm.}
}
\value{
MAF file annotated with clonality estimates for each mutation.
}
\description{
Based on FACETS data, infer cancer-cell fraction (CCF) for somatic mutations in a sample.
Output columns:
}
