% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf-annotate-maf.R
\name{ccf_annotate_maf_legacy}
\alias{ccf_annotate_maf_legacy}
\title{Estimate CCFs of somatic mutations from cncf.txt file.}
\usage{
ccf_annotate_maf_legacy(
  maf,
  cncf_txt_file,
  purity,
  algorithm = c("em", "cncf")
)
}
\arguments{
\item{maf}{Input MAF file.}

\item{cncf_txt_file}{.cncf.txt file created with legacy output of facets-suite.}

\item{purity}{Sample purity estimate.}

\item{algorithm}{Choice between FACETS \code{em} and \code{cncf} algorithm.}
}
\value{
MAF file annotated with clonality estimates for each mutation, where the following column prefixes are used:
\itemize{
  \item{\code{ccf_Mcopies*}:} {Inferred CCF if mutation is on the major allele.}
  \item{\code{ccf_1copy*}:} {Inferred CCF if mutation exists in one copy.}
  \item{\code{ccf_expected_copies*}:} {Inferred CCF if mutation exists in number of copies expected from observed VAF and local ploidy.}
}
}
\description{
Based on FACETS data, infer cancer-cell fraction (CCF) for somatic mutations in a sample.
}
