% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots_ggplot2.R
\name{close.up}
\alias{close.up}
\title{don't know}
\usage{
\method{close}{up}(out, fit, chrom.range = NULL, method = NA,
  gene.name = NULL, lend = "butt", bed.path = NULL,
  subset.snps = FALSE, plotX = FALSE, ...)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{chrom.range}{numeric first argument}

\item{method}{numeric first argument}

\item{gene.name}{numeric first argument}

\item{lend}{numeric first argument}

\item{bed.path}{numeric first argument}

\item{subset.snps}{numeric first argument}

\item{plotX}{numeric first argument}

\item{theme}{numeric first argument}
}
\value{
don't know
}
\description{
Need to add this functionality so it can be callled by the wrapper, doFacets.R etc.
}
