% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLevel.R
\name{convert_gene_level_calls_to_matrix_portal}
\alias{convert_gene_level_calls_to_matrix_portal}
\alias{get_gene_level_calls}
\title{don't know}
\usage{
get_gene_level_calls(cncf_files, gene_targets, WGD_threshold = 0.5,
  amp_threshold = 5, max_seg_length = 2.5e+07, min_cf = 0.6,
  mean_chrom_threshold = 0, fun.rename = function(filename) {    
  filename })

convert_gene_level_calls_to_matrix_portal(gene_level_calls)
}
\arguments{
\item{cncf_files}{numeric first argument}

\item{gene_targets}{numeric first argument}

\item{WGD_threshold}{numeric first argument}

\item{amp_threshold}{numeric first argument}

\item{max_seg_length}{numeric first argument}

\item{min_cf}{numeric first argument}

\item{mean_chrom_threshold}{numeric first argument}

\item{fun.rename}{numeric first argument}

\item{gene_level_calls}{numeric first argument}
}
\value{
don't know

don't know
}
\description{
don't know

don't know
}
