% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots_ggplot2.R
\name{copy.number.log.ratio}
\alias{copy.number.log.ratio}
\title{don't know}
\usage{
copy.number.log.ratio(out, fit, load.genome = FALSE, gene.pos = NULL,
  col.1 = "#0080FF", col.2 = "#4CC4FF", sample.num = NULL,
  lend = "butt", theme = "bw", subset.indices = NULL,
  plotX = FALSE)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{load.genome}{numeric first argument}

\item{gene.pos}{numeric first argument}

\item{col.1}{numeric first argument}

\item{col.2}{numeric first argument}

\item{sample.num}{numeric first argument}

\item{lend}{numeric first argument}

\item{theme}{numeric first argument}

\item{subset.indices}{numeric first argument}

\item{plotX}{numeric first argument}
}
\value{
don't know
}
\description{
temporary function
}
