% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno_fast.R
\name{estimated_af_life_history}
\alias{estimated_af_life_history}
\title{don't know}
\usage{
estimated_af_life_history(purity, ns, nw, m, M, copies = 1,
  limit = TRUE)
}
\arguments{
\item{purity}{numeric first argument}

\item{ns}{numeric first argument}

\item{nw}{numeric first argument}

\item{m}{numeric first argument}

\item{M}{numeric first argument}

\item{copies}{numeric first argument}

\item{limit}{boolean first argument}
}
\value{
don't know
}
\description{
don't know
}
