% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFacets.R
\name{extract_six_column_counts_matrix}
\alias{extract_six_column_counts_matrix}
\title{don't know}
\usage{
extract_six_column_counts_matrix(COUNTS_FILE)
}
\arguments{
\item{COUNTS_FILE}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
