\name{facets-suite-package}
\alias{facets-suite-package}
\alias{facets-suite}
\docType{package}
\title{
\packageTitle{facets-suite}
}
\description{
\packageDescription{facets-suite}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{facets-suite}
\packageIndices{facets-suite}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{facets-suite}

Maintainer: \packageMaintainer{facets-suite}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
