% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetsSuite.R
\docType{package}
\name{facetsSuite}
\alias{facetsSuite}
\alias{facetsSuite-package}
\title{\code{facetsSuite} package}
\description{
Functions to run and parse output from \href{https://github.com/mskcc/facets}{FACETS}.
}
