% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-igv-seg.R
\name{format_igv_seg}
\alias{format_igv_seg}
\title{Format `.seg` file}
\usage{
format_igv_seg(facets_output, sample_id, normalize = TRUE)
}
\arguments{
\item{facets_output}{Facets output, from `run_facets`.}

\item{sample_id}{Sample name.}

\item{normalize}{Adjust copy-number log-ratio by dipLogR.}
}
\value{
Segmentation output formatted for IGV.
}
\description{
Format Facets output for viewing in IGV.
}
