% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno_fast.R
\name{fread_rbind}
\alias{fread_rbind}
\title{don't know}
\usage{
fread_rbind(filenames, fn = fread)
}
\arguments{
\item{filenames}{numeric first argument}

\item{fn}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
