% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-level-changes.R
\name{gene_level_changes}
\alias{gene_level_changes}
\title{Get gene-level changes in copy number from FACETS output.}
\usage{
gene_level_changes(facets_output, genome = c("hg19", "hg38"),
  algorithm = c("em", "cncf"))
}
\arguments{
\item{facets_output}{Full FACETS output from \code{run_facets}.}

\item{genome}{Genome build.}

\item{algorithm}{Choice between FACETS \code{em} and \code{cncf} algorithm.}
}
\value{
\code{data.frame} for all genes mapping onto a segment in the output segmentation, with the columns:
\itemize{
    \item{\code{median_cnlr_seg}:} {Median copy-number log-ratio for segment.}
    \item{\code{genes_on_seg}:} {Protein-coding genes on segment.}
    \item{\code{gene_snps}:} {Median copy-number log-ratio for segment.}
    \item{\code{gene_het_snps}:} {Median copy-number log-ratio for segment.}
    \item{\code{spans_segs}:} {Boolean indicating whether the gene boundaries map onto different segments.}
    \item{\code{cn_state}:} {Label for coyp-number configuration.}
    \item{\code{mean_cnlr}:} {Mean copy-number log-ratio for SNPs in gene.}
    \item{\code{pval}:} {P-value from Z-test, as described above.}
    \item{\code{fold_change}:} {Fold change corresponding to the comparison against the dipLogR baseline.}
}
}
\description{
This maps protein-coding genes onto copy-number segmentation from FACETS output derives the copy-number status from that, based on the integer copy number.
Additionally, the copy-number log-ratio of the SNPs falling insided the gene boundaries are used to perform a Z-test against the dipLogR baseline.
}
\examples{
gene_level_changes(test_facets_output, 'hg38', 'em')
}
