% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_level_merge.R
\name{gene_level_merge_function}
\alias{gene_level_merge_function}
\title{don't know}
\usage{
gene_level_merge_function(hisens, purity, outfile)
}
\arguments{
\item{hisens}{character FACETS hisens gene-level calls}

\item{purity}{character FACETS purity gene-level calls}

\item{outfile}{character Output filename}
}
\value{
don't know
}
\description{
don't know
}
