% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots_ggplot2.R
\name{get.cumulative.chr.maploc}
\alias{get.cumulative.chr.maploc}
\title{don't know}
\usage{
get.cumulative.chr.maploc(mat, load.genome = FALSE)
}
\arguments{
\item{mat}{numeric first argument}

\item{load.genome}{boolean first argument}
}
\value{
don't know
}
\description{
temporary function
}
