% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots_ggplot2.R
\name{get.gene.pos}
\alias{get.gene.pos}
\title{don't know}
\usage{
get.gene.pos(hugo.symbol, my.path = system.file("extdata",
  "Homo_sapiens.GRCh37.75.canonical_exons.bed", package = "facets.suite"),
  load.genome = FALSE)
}
\arguments{
\item{hugo.symbol}{numeric first argument}

\item{my.path}{numeric first argument}

\item{load.genome}{numeric first argument}
}
\value{
don't know
}
\description{
temporary function
}
