% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots_ggplot2.R
\name{integer.copy.number}
\alias{integer.copy.number}
\title{don't know}
\usage{
integer.copy.number(out, fit, method = c("cncf", "em"),
  load.genome = FALSE, gene.pos = NULL, main = "", lend = "butt",
  theme = "bw", plotX = FALSE, ...)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{method}{numeric first argument}

\item{load.genome}{numeric first argument}

\item{gene.pos}{numeric first argument}

\item{main}{numeric first argument}

\item{lend}{numeric first argument}

\item{theme}{numeric first argument}

\item{plotX}{numeric first argument}
}
\value{
don't know
}
\description{
temporary function
}
