% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafAnno.R, R/mafAnno_fast.R,
%   R/test_facets_fit.R
\name{main}
\alias{main}
\title{don't know}
\usage{
main(maf, facets, plot = FALSE)

main(maf, facets, plot = FALSE)

main(maf, facets, plot = FALSE)
}
\arguments{
\item{maf}{numeric first argument}

\item{facets}{numeric first argument}

\item{plot}{numeric first argument}

\item{facets_file}{numeric first argument}

\item{file_type}{numeric first argument}

\item{maf}{numeric first argument}

\item{cncf_files}{numeric first argument}

\item{purity_fit_filename}{numeric first argument}

\item{maf}{numeric first argument}
}
\value{
don't know

don't know

don't know
}
\description{
don't know

temporary function

don't know
}
