% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{makeLegacySegObject}
\alias{makeLegacySegObject}
\title{don't know}
\usage{
makeLegacySegObject(out, sampID)
}
\arguments{
\item{out}{numeric first argument}

\item{sampID}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
