% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_level_merge.R
\name{merge_calls}
\alias{merge_calls}
\title{don't know}
\usage{
merge_calls(hisens, purity)
}
\arguments{
\item{hisens}{numeric first argument}

\item{purity}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
