% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_normal_depth.R
\name{normalize_facets_depth}
\alias{normalize_facets_depth}
\title{don't know}
\usage{
normalize_facets_depth(filename, out_filename)
}
\arguments{
\item{filename}{numeric first argument}

\item{out_filename}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
