% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots_ggplot2.R
\name{plot.facets.all.output}
\alias{plot.facets.all.output}
\title{don't know}
\usage{
\method{plot}{facets.all.output}(out, fit, w = 850, h = 1100,
  type = "png", load.genome = FALSE, main = "", plotname = "test",
  gene.name = NULL, lend = "butt", em.plot = FALSE,
  subset.snps = FALSE)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{w}{numeric first argument}

\item{h}{numeric first argument}

\item{type}{numeric first argument}

\item{load.genome}{numeric first argument}

\item{main}{numeric first argument}

\item{plotname}{numeric first argument}

\item{gene.name}{numeric first argument}

\item{lend}{numeric first argument}

\item{em.plot}{numeric first argument}

\item{subset.snps}{numeric first argument}
}
\value{
don't know
}
\description{
Standard facets output plot
}
