% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-facets.R
\name{plot_facets}
\alias{plot_facets}
\alias{cnlr_plot}
\alias{valor_plot}
\alias{cf_plot}
\alias{icn_plot}
\title{Plot FACETS output}
\usage{
cnlr_plot(facets_data, cols = c("#0080FF", "#4CC4FF"),
  subset_indices = NULL, plotX = FALSE, genome = c("hg19", "hg18",
  "hg38"), gene_pos = NULL, adjust_diplogr = TRUE)

valor_plot(facets_data, cols = c("#0080FF", "#4CC4FF"),
  subset_indices = NULL, plotX = FALSE, genome = c("hg19", "hg18",
  "hg38"), gene_pos = NULL)

cf_plot(facets_data, method = c("em", "cncf"), plotX = FALSE,
  genome = c("hg19", "hg18", "hg38"))

icn_plot(facets_data, method = c("em", "cncf"), plotX = FALSE,
  genome = c("hg19", "hg18", "hg38"), gene_pos = NULL)
}
\arguments{
\item{facets_data}{Output object from \code{run_facets}.}

\item{cols}{Vector of two colors for alternating chromosomes.}

\item{subset_indices}{???.}

\item{plotX}{Plot chromosome X.}

\item{genome}{Genome build.}

\item{gene_pos}{Highlight gene.}

\item{adjust_diplogr}{Normalize by sample dipLogR.}

\item{method}{When available, choose between plotting solution from \code{em} or \code{cncf} algorithm.}
}
\value{
\code{ggplot2} objects.
}
\description{
Generate the following plots:
\itemize{
 \item{Log-ratio:} {Copy-number segmentation based on tumor-normal read coverage comparison.}
 \item{Allelic imbalance:} {Allelic imbalance based on somatic changes in zygosity of heterozygous SNPs in the normal.}
 \item{Integer copy number:} {Inference of the major and minor copy-number states based on the tumor-normal log ratio and allelic content.}
 \item{Cellular fraction:} {Estimate of fraction of cells in sample harboring each segment.}
}
}
