% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-segmentation.R
\name{plot_segmentation}
\alias{plot_segmentation}
\title{Plot segmentation profile}
\usage{
plot_segmentation(segs, plotX = FALSE, sample_order = NULL,
  cap_log_ratios = TRUE, colors = c("darkblue", "white", "red"),
  return_object = FALSE)
}
\arguments{
\item{segs}{FACETS segment output, IGV formatted.}

\item{plotX}{If \code{TRUE}, includes chromosome X.}

\item{sample_order}{Manual order of samples.}

\item{cap_log_ratios}{Cap log-ratios at the absolute value.}

\item{colors}{Vector of three colors, giving the low-, mid- and high-point of the color scale.}

\item{return_object}{If \code{TRUE}, returns \code{ggplot2} object instead of printing plot.}
}
\value{
Output plots in viewer, unless \code{return_object} is used and \code{ggplot2} objects are returned.
}
\description{
Creates an IGV-like graphical representation of the copy-number segments across the samples in a segmentation object, as output by \code{run_facets}.
}
