% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets_qc.R, R/test_facets_fit.R
\name{plot_vaf_by_cn_state}
\alias{plot_vaf_by_cn_state}
\title{don't know}
\usage{
plot_vaf_by_cn_state(maf, sample, purity, wgd = FALSE)

plot_vaf_by_cn_state(maf, sample, purity, wgd = FALSE)
}
\arguments{
\item{maf}{numeric first argument}

\item{sample}{numeric first argument}

\item{purity}{numeric first argument}

\item{wgd}{numeric first argument}

\item{maf}{numeric first argument}

\item{sample}{numeric first argument}

\item{wgd}{numeric first argument}
}
\value{
don't know

don't know
}
\description{
don't know

don't know
}
