% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plotSample3}
\alias{plotSample3}
\title{don't know}
\usage{
plotSample3(x, pch = ".", cex = 2, label = "NA")
}
\arguments{
\item{x}{numeric first argument}

\item{pch}{numeric first argument}

\item{cex}{numeric first argument}

\item{label}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
