% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plotSampleCNCF2}
\alias{plotSampleCNCF2}
\title{don't know}
\usage{
plotSampleCNCF2(out, fit)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
