% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFacets.R
\name{print_run_details}
\alias{print_run_details}
\title{don't know}
\usage{
print_run_details(out, fit, COUNTS_FILE, TAG, DIRECTORY, CVAL, DIPLOGR,
  NDEPTH, SNP_NBHD, MIN_NHET, GENOME, GGPLOT, SINGLE_CHROM, SEED,
  RLIB_PATH, RLIB_VERSION, GIVE_PCVAL, unmatched, TUMOR_ID)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{COUNTS_FILE}{numeric first argument}

\item{TAG}{numeric first argument}

\item{DIRECTORY}{numeric first argument}

\item{CVAL}{numeric first argument}

\item{DIPLOGR}{numeric first argument}

\item{NDEPTH}{numeric first argument}

\item{SNP_NBHD}{numeric first argument}

\item{MIN_NHET}{numeric first argument}

\item{GENOME}{numeric first argument}

\item{GGPLOT}{numeric first argument}

\item{SINGLE_CHROM}{numeric first argument}

\item{SEED}{numeric first argument}

\item{RLIB_PATH}{numeric first argument}

\item{RLIB_VERSION}{numeric first argument}

\item{GIVE_PCVAL}{numeric first argument}

\item{unmatched}{numeric first argument}

\item{TUMOR_ID}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
