% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality-check.R
\name{quality_check}
\alias{quality_check}
\title{Sample QC}
\usage{
quality_check(facets_output, maf = NULL, genome = c("hg19", "hg18",
  "hg38"), algorithm = c("em", "cncf"))
}
\arguments{
\item{facets_output}{}

\item{maf}{}

\item{algorithm}{}
}
\description{
Generate QC metrics for sample. Can use mutation calls in MAF file.
}
