% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlots_ggplot2.R
\name{random.subset.snps}
\alias{random.subset.snps}
\title{don't know}
\usage{
random.subset.snps(jointseg, by_factor = 5)
}
\arguments{
\item{jointseg}{numeric first argument}

\item{by_factor}{numeric first argument}
}
\value{
don't know
}
\description{
Subset SNPs for plots that can be handled by e.g. Illustrator, defaults to 5-fold downsampling at present
}
