% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-snp-matrix.R
\name{read_snp_matrix}
\alias{read_snp_matrix}
\title{Read counts file}
\source{
\code{snp-pileup is part of} \url{www.github.com/mskcc/facets}
}
\usage{
read_snp_matrix(input_file, err.thresh = 10, del.thresh = 10)
}
\arguments{
\item{input_file}{Path to input file.}

\item{err.thresh}{Threshold for errors at locus.}

\item{del.thresh}{Threshold for deletions at locus.}
}
\value{
Count matrix.
}
\description{
Reads a gzipped output file from \code{snp-pileup}.
}
