% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-snp-matrix.R
\name{read_snp_matrix_facets2n}
\alias{read_snp_matrix_facets2n}
\title{#' Read in the snp-pileup generated SNP read count matrix file for facets2n processing}
\source{
\code{snp-pileup} is part of \href{www.github.com/mskcc/facets}{FACETS}.
}
\usage{
read_snp_matrix_facets2n(
  filename,
  err.thresh = Inf,
  del.thresh = Inf,
  perl.pileup = FALSE,
  MandUnormal = FALSE,
  spanT = 0.2,
  spanA = 0.2,
  spanX = 0.2,
  gbuild = "hg19",
  ReferencePileupFile = NULL,
  ReferenceLoessFile = NULL,
  MinOverlap = 0.9,
  useMatchedX = FALSE,
  refX = FALSE,
  unmatched = FALSE,
  donorCounts = FALSE
)
}
\arguments{
\item{filename}{counts file from snp-pileup}

\item{err.thresh}{(numeric) Error threshold to be used to filter snp-pileup data frame.}

\item{del.thresh}{(numeric) Deletion threshold to be used to filter snp-pileup data frame.}

\item{perl.pileup}{(logical) Is the pileup data generated using perl pileup tool?}

\item{MandUnormal}{(logical) Is CNLR analysis to be peformed using unmatched reference normals?}

\item{spanT}{(numeric) Default span value to be used for loess normalization in tumor sample.}

\item{spanA}{(numeric) Default span value to be used for loess normalization across autosomal chromosomes in the normal sample.}

\item{spanX}{(numeric) Default span value to be used for loess normalization in Chr X in the normal sample.}

\item{gbuild}{(character) Genome build (Default: hg19).}

\item{ReferencePileupFile}{(character) Filepath to an optional snp-pileup generated pileup data of one or more reference normals.}

\item{ReferenceLoessFile}{(character) Filepath to an optional loess data, generated using the facets2n package, of one or more reference normals. The number of normals in this data should match that in the ReferencePileupFile, and should be in the same order.}

\item{MinOverlap}{(numeric) Mininum overlap fraction of loci between a tumor pileup and reference pileup data.}

\item{useMatchedX}{(logical) Is the matched normal to be used for ChrX normalization?}

\item{refX}{(logical) Use sex matched reference normal for chrX normalization}

\item{unmatched}{(logical) is the tumor being analyzed unmatched}

\item{donorCounts}{snp read count matrix for donor sample(s). Required columns: Chromosome Position Ref Alt and for each donor sample,i: RefDonoriR RefDonoriA RefDonoriE RefDonoriD RefDonoriDP}

\item{skip}{(character) Skip n number of lines in the input file.}
}
\value{
A dataframe of pileup depth values for Tumor and Matched Normal if MandUnormal is FALSE. Else, a list of data frame with pileup depth values of Tumor, matched Normal, and a best unmatched normal, and the associated span values.
}
\description{
#' Read in the snp-pileup generated SNP read count matrix file for facets2n processing
}
