% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFacets.R
\name{readSnpMatrix2}
\alias{readSnpMatrix2}
\title{don't know}
\usage{
readSnpMatrix2(pileup, err.thresh = Inf, del.thresh = Inf)
}
\arguments{
\item{pileup}{numeric first argument}

\item{err.thresh}{numeric first argument}

\item{del.thresh}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
