% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFacets.R
\name{results_figure}
\alias{results_figure}
\title{don't know}
\usage{
results_figure(out, fit, DIRECTORY, TAG, CVAL, GGPLOT, SINGLE_CHROM,
  GIVE_PCVAL, EM_PLOT = FALSE)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{DIRECTORY}{numeric first argument}

\item{TAG}{numeric first argument}

\item{CVAL}{numeric first argument}

\item{GGPLOT}{numeric first argument}

\item{SINGLE_CHROM}{numeric first argument}

\item{GIVE_PCVAL}{numeric first argument}

\item{EM_PLOT}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
