% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-facets.R
\name{run_facets}
\alias{run_facets}
\title{Run FACETS}
\usage{
run_facets(
  read_counts,
  cval = 100,
  dipLogR = NULL,
  ndepth = 25,
  snp_nbhd = 250,
  min_nhet = 10,
  genome = c("hg18", "hg19", "hg38", "mm9", "mm10"),
  seed = 100,
  facets_lib_path = "",
  facets2n_lib_path = "",
  MandUnormal = FALSE,
  useMatchedX = FALSE,
  referencePileup = NULL,
  referenceLoess = NULL,
  unmatched = FALSE
)
}
\arguments{
\item{read_counts}{Read counts object, generated by `snp-pileup`.}

\item{cval}{Segmentation parameter, higher values for higher sensitivity.}

\item{dipLogR}{Manual dipLogR value, if left empty `facets` finds the most likely sample baseline.}

\item{ndepth}{Minimum depth in normal to retain SNP, see `facets` help.}

\item{snp_nbhd}{Minimum basepair distance for SNPs, see `facets` help.}

\item{min_nhet}{Minimum number of heterozygous SNPs on segment required for clustering, see `facets` help.}

\item{genome}{Genome build.}

\item{seed}{Seed value for random number generation, set to enable full reproducibility.}

\item{facets_lib_path}{path to facets R library}

\item{facets2n_lib_path}{path to facets2n R library}

\item{MandUnormal}{(logical) facets2n option: Is CNLR analysis to be peformed using unmatched reference normals?}

\item{useMatchedX}{(logical) facets2n option: Force matched normal to be used for ChrX normalization?}

\item{referencePileup}{facets2n option: Filepath to an optional snp-pileup generated pileup data of one or more reference normals.}

\item{referenceLoess}{facets2n option: Filepath to an optional loess data, generated using the facets2n package, of one or more reference normals. The number of normals in this data should match that in the ReferencePileupFile, and should be in the same order.}

\item{unmatched}{(logical) facets2n option: unmatched, use tumor for determination of heterzygous SNPs}
}
\value{
A list object containing the following items. See \href{www.github.com/mskcc/facets}{FACETS documentation} for more details:
\itemize{
      \item{\code{snps}:} {SNPs used for copy-number segmentation, \code{het==1} for heterozygous loci.}
      \item{\code{segs}:} {Inferred copy-number segmentation.}
      \item{\code{purity}:} {Inferred sample purity, i.e. fraction of tumor cells of the total cellular population.}
      \item{\code{ploidy}:} {Inferred sample ploidy.}
      \item{\code{dipLogR}:} {Inferred dipLogR, the sample-specific baseline corresponding to the diploid state.}
      \item{\code{alBalLogR}:} {Alternative dipLogR value(s) at which a balanced solution was found.}
      \item{\code{flags}:} {Warning flags from the naïve segmentation algorithm.}
      \item{\code{em_flags}:} {Warning flags from the expectation-maximization segmentation algorithm.}
      \item{\code{loglik}:} {Log-likelihood value of the fitted model.}
}
}
\description{
Runs FACETS on an input count file, with specified parameter settings.
}
\examples{
\dontrun{
library(pctGCdata)
run_facets(test_read_counts, cval = 500, genome = 'hg38')
}

}
