% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{runEM2}
\alias{runEM2}
\title{don't know}
\usage{
runEM2(d, matched.normal = TRUE)
}
\arguments{
\item{d}{numeric first argument}

\item{matched.normal}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
