% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFacets.R
\name{seg_figure}
\alias{seg_figure}
\title{don't know}
\usage{
seg_figure(out, DIRECTORY, TAG, chromLevels, CVAL)
}
\arguments{
\item{out}{numeric first argument}

\item{DIRECTORY}{numeric first argument}

\item{TAG}{numeric first argument}

\item{chromLevels}{numeric first argument}

\item{CVAL}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
