% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFacets.R
\name{select_genome}
\alias{select_genome}
\title{don't know}
\usage{
select_genome(GENOME, SINGLE_CHROM)
}
\arguments{
\item{GENOME}{numeric first argument}

\item{SINGLE_CHROM}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
