% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\docType{data}
\name{test_data}
\alias{test_data}
\alias{test_read_counts}
\title{Test data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 21608713 rows and 6 columns.}
\source{
\url{portal.gdc.cancer.gov}
}
\usage{
data(test_read_counts); data(test_facets_output); data(test_maf)
}
\description{
Data for user and development testing, based on TCGA sample TCGA-06-0154.
All data from GDC's data portal, processed as described below.
}
\keyword{datasets}
