% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\docType{data}
\name{test_facets_output}
\alias{test_facets_output}
\title{Test FACETS output
#' @format Output from \code{run-facets}.}
\format{An object of class \code{list} of length 9.}
\source{
\url{portal.gdc.cancer.gov}
}
\usage{
test_facets_output
}
\description{
Test FACETS output
#' @format Output from \code{run-facets}.
}
\keyword{datasets}
