% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\docType{data}
\name{test_maf}
\alias{test_maf}
\title{Test MAF}
\format{Mutation annotation format (MAF) with standard columns.}
\source{
\url{portal.gdc.cancer.gov}
}
\usage{
test_maf
}
\description{
Test MAF
}
\keyword{datasets}
