% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\docType{data}
\name{test_read_counts}
\alias{test_read_counts}
\title{Test read counts}
\format{Output from \code{snp-pileup}.}
\source{
\url{portal.gdc.cancer.gov}
}
\usage{
test_read_counts
}
\description{
Test read counts
}
\keyword{datasets}
