% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doFacets.R
\name{write_output}
\alias{write_output}
\title{don't know}
\usage{
write_output(out, fit, DIRECTORY, TAG, TUMOR_ID)
}
\arguments{
\item{out}{numeric first argument}

\item{fit}{numeric first argument}

\item{DIRECTORY}{numeric first argument}

\item{TAG}{numeric first argument}

\item{TUMOR_ID}{numeric first argument}
}
\value{
don't know
}
\description{
don't know
}
