% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-number-scores.R
\name{copy_number_scores}
\alias{copy_number_scores}
\alias{calculate_fcna}
\title{Calculate copy-number based scores}
\usage{
calculate_fcna(segs, ploidy, genome = c("hg18", "hg19", "grch38"),
  algorithm = c("em", "cncf"))
}
\arguments{
\item{algorithm}{}
}
\description{
Calculate copy-number based scores
}
