% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-seg.R
\name{format_seg}
\alias{format_seg}
\title{Format .seg file}
\usage{
format_seg(facets_output, sample_id, normalize = TRUE)
}
\arguments{
\item{sample_id}{Sample name.}

\item{out}{Facets out objet.}
}
\value{
Segmentation output formatted for IGV.
}
\description{
Format Facets output for viewing in IGV.
}
