% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-facets.R
\name{plot_facets}
\alias{plot_facets}
\alias{cnlr_plot}
\alias{valor_plot}
\alias{cf_plot}
\alias{icn_plot}
\alias{closeup_plot}
\title{Plot FACETS output}
\usage{
cnlr_plot(facets_data, colors = c("#0080FF", "#4CC4FF"), plotX = FALSE,
  genome = c("hg19", "hg18", "hg38"), highlight_gene = NULL,
  adjust_diplogr = TRUE, subset_snps = NULL, return_object = FALSE)

valor_plot(facets_data, colors = c("#0080FF", "#4CC4FF"),
  plotX = FALSE, genome = c("hg19", "hg18", "hg38"),
  highlight_gene = NULL, subset_snps = NULL, return_object = FALSE)

cf_plot(facets_data, method = c("em", "cncf"), plotX = FALSE,
  genome = c("hg19", "hg18", "hg38"), return_object = FALSE)

icn_plot(facets_data, method = c("em", "cncf"), plotX = FALSE,
  genome = c("hg19", "hg18", "hg38"), highlight_gene = NULL,
  return_object = FALSE)

closeup_plot(facets_data, highlight_gene = NULL, plot_chroms = NULL,
  method = c("em", "cncf"), genome = c("hg19", "hg18", "hg38"),
  return_object = FALSE)
}
\arguments{
\item{facets_data}{Output object from \code{run_facets}.}

\item{colors}{Vector of two colors for alternating chromosomes.}

\item{plotX}{If \code{TRUE}, includes chromosome X.}

\item{genome}{Genome build.}

\item{highlight_gene}{Highlight gene(s), provide gene symbol or mapped positions (internally).}

\item{adjust_diplogr}{Normalize by sample dipLogR.}

\item{subset_snps}{Subset the SNP profile to reduce weight of plotting, supply a factor by which to reduce or \code{TRUE} for default.}

\item{return_object}{If \code{TRUE}, returns \code{ggplot2} object instead of printing plot.}

\item{method}{When available, choose between plotting solution from \code{em} or \code{cncf} algorithm.}

\item{plot_chroms}{Chromosomes to plot when using \code{closeup_plot}.}
}
\value{
\code{ggplot2} objects, see input parameter \code{return_object}.
}
\description{
Generate the following plots:
\itemize{
 \item{Log-ratio:} {Copy-number segmentation based on tumor-normal read coverage comparison.}
 \item{Allelic imbalance:} {Allelic imbalance based on somatic changes in zygosity of heterozygous SNPs in the normal.}
 \item{Integer copy number:} {Inference of the major and minor copy-number states based on the tumor-normal log ratio and allelic content.}
 \item{Cellular fraction:} {Estimate of fraction of cells in sample harboring each segment.}
}
}
