% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-snp-matrix.R
\name{read_snp_matrix}
\alias{read_snp_matrix}
\title{Read counts file}
\source{
\code{snp-pileup} is part of \href{www.github.com/mskcc/facets}{FACETS}.
}
\usage{
read_snp_matrix(input_file, err.thresh = 10, del.thresh = 10)
}
\arguments{
\item{input_file}{Path to input file.}

\item{err.thresh}{Threshold for errors at locus.}

\item{del.thresh}{Threshold for deletions at locus.}
}
\value{
A SNP count matrix, with the following columns:
\itemize{
      \item{\code{Chromosome}:} {SNPs used for copy-number segmentation, \code{het==1} for heterozygous loci.}
      \item{\code{Position}:} {Inferred copy-number segmentation.}
      \item{\code{NOR.DP}:} {Total read depth in normal sample.}
      \item{\code{TUM.DP}:} {Total read depth in tumor sample.}
      \item{\code{NOR.RD}:} {Reference allele read depth in normal sample.}
      \item{\code{TUM.RD}:} {Reference allele read depth in tumor sample.}
}
}
\description{
Reads a gzipped output file from \code{snp-pileup}.
}
