% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-facets.R
\name{run_facets}
\alias{run_facets}
\title{Run FACETS}
\arguments{
\item{counts}{Counts object.}
}
\value{
A list object with \code{out} and \code{fit} objects from Facets run.
}
\description{
Runs FACETS on an input count file, with specified parameter settings.
}
\examples{
facets_closeup('path_to_sample.Rdata', 'BRCA1', 17)

}
