% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-facets.R
\name{run_facets}
\alias{run_facets}
\title{Run FACETS}
\usage{
run_facets(read_counts, cval = 100, diplogr = NULL, ndepth = 35,
  snp_nbhd = 250, min_nhet = 15, genome = c("hg18", "hg19", "hg38",
  "mm9", "mm10"), seed = 100)
}
\arguments{
\item{read_counts}{Read counts object, generated by `snp-pileup`.}

\item{cval}{Segmentation parameter, higher values for higher sensitivity.}

\item{diplogr}{Manual dipLogR value, if left empty `facets` finds the most likely sample baseline.}

\item{ndepth}{Minimum depth in normal to retain SNP, see `facets` help.}

\item{snp_nbhd}{Minimum basepair distance for SNPs, see `facets` help.}

\item{min_nhet}{Minimum number of heterozygous SNPs on segment required for clustering, see `facets` help.}

\item{genome}{Genome build.}

\item{seed}{Seed value for random number generation, set to enable full reproducibility.}
}
\value{
A list object containing the following items. See \href{www.github.com/mskcc/facets}{FACETS documentation} for more details:
\itemize{
      \item{\code{snps}:} {SNPs used for copy-number segmentation, \code{het==1} for heterozygous loci.}
      \item{\code{segs}:} {Inferred copy-number segmentation.}
      \item{\code{purity}:} {Inferred sample purity, i.e. fraction of tumor cells of the total cellular population.}
      \item{\code{ploidy}:} {Inferred sample ploidy.}
      \item{\code{diplogr}:} {Inferred dipLogR, the sample-specific baseline corresponding to the diploid state.}
      \item{\code{alballogr}:} {Alternative dipLogR value(s) at which a balanced solution was found.}
      \item{\code{flags}:} {Warning flags from the naïve segmentation algorithm.}
      \item{\code{em_flags}:} {Warning flags from the expectation-maximization segmentation algorithm.}
      \item{\code{loglik}:} {Log-likelihood value of the fitted model.}
}
}
\description{
Runs FACETS on an input count file, with specified parameter settings.
}
\examples{
\dontrun{
library(pctGCdata)
run_facets(test_read_counts, cval = 500, genome = 'hg38')
}

}
