% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-facets.R
\name{run_facets}
\alias{run_facets}
\title{Run FACETS}
\usage{
run_facets(read_counts, cval = 100, diplogr = NULL, ndepth = 35,
  snp_nbhd = 250, min_nhet = 15, genome = c("hg18", "hg19", "hg38",
  "mm9", "mm10"), seed = 100, rlib_path = NULL)
}
\arguments{
\item{read_counts}{Read counts object, generated by `snp-pileup`.}

\item{cval}{Segmentation parameter, higher values for higher sensitivity.}

\item{diplogr}{Manual dipLogR value, if left empty `facets` finds the most likely sample baseline.}

\item{ndepth}{Minimum depth in normal to retain SNP, see `facets` help.}

\item{snp_nbhd}{Minimum basepair distance for SNPs, see `facets` help.}

\item{min_nhet}{Minimum number of heterozygous SNPs on segment required for clustering, see `facets` help.}

\item{genome}{Genome build.}

\item{seed}{Seed value for random number generation, set to enable full reproducibility.}

\item{rlib_path}{}
}
\value{
A list object with \code{out} and \code{fit} objects from Facets run.
}
\description{
Runs FACETS on an input count file, with specified parameter settings.
}
\examples{
library(pctGCdata)
run_facets(test_read_counts, cval = 500, genome = 'hg38')

}
